backend/requirements.txt
fastapi
uvicorn
qdrant-client
google-generativeai
python-dotenv
asyncpg
databases
passlib[bcrypt]
pyjwt
sqlalchemy
psycopg2-binary
python-jose
pydantic-settings
